ENH: Working SRR pipeline and clean interfaces
